+++ src/Testcontainers/Clients/DockerSystemOperations.cs
@@ -3,6 +3,7 @@ namespace DotNet.Testcontainers.Clients
   using System;
   using System.Threading;
   using System.Threading.Tasks;
+  using Docker.DotNet.Models;
   using DotNet.Testcontainers.Configurations;
   using Microsoft.Extensions.Logging;
 
@@ -18,6 +19,11 @@ namespace DotNet.Testcontainers.Clients
     {
       return (await this.Docker.System.GetSystemInfoAsync(ct)
         .ConfigureAwait(false)).OperatingSystem.Contains("Windows");
+    }
+
+    public Task<VersionResponse> GetVersion(CancellationToken ct = default)
+    {
+      return this.Docker.System.GetVersionAsync(ct);
     }
   }
 }